+++ tests/test_backends.py
@@ -386,6 +386,49 @@ def test_jvp_pytree(backend):
     np.testing.assert_allclose(g["a"], np.ones([2]), atol=1e-5)
 
 
+@pytest.mark.parametrize("backend", [lf("tfb"), lf("jaxb")])
+def test_jacfwd(backend):
+    def f(x):
+        return x ** 2
+
+    x = tc.backend.ones([3])
+    jacf = tc.backend.jacfwd(f)
+    np.testing.assert_allclose(jacf(x), 2 * np.eye(3), atol=1e-5)
+
+    def f2(x):
+        return x ** 2, x
+
+    jacf2 = tc.backend.jacfwd(f2)
+    np.testing.assert_allclose(jacf2(x)[1], np.eye(3), atol=1e-5)
+    np.testing.assert_allclose(jacf2(x)[0], 2 * np.eye(3), atol=1e-5)
+
+    def f3(x, y):
+        return x + y ** 2
+
+    jacf3 = tc.backend.jacfwd(f3, argnums=(0, 1))
+    np.testing.assert_allclose(jacf3(x, x)[1], 2 * np.eye(3), atol=1e-5)
+
+    def f4(x, y):
+        return x ** 2, y
+
+    jacf4 = tc.backend.jacfwd(f4, argnums=(0, 1))
+    np.testing.assert_allclose(jacf4(x, x)[1][1], np.eye(3), atol=1e-5)
+    np.testing.assert_allclose(jacf4(x, x)[0][1], np.zeros([3, 3]), atol=1e-5)
+
+
+@pytest.mark.parametrize("backend", [lf("tfb"), lf("jaxb")])
+def test_jacfwd_tall(backend):
+    h = tc.backend.ones([5, 3])
+
+    def f(x):
+        x = tc.backend.reshape(x, [-1, 1])
+        return tc.backend.reshape(h @ x, [-1])
+
+    x = tc.backend.ones([3])
+    jacf = tc.backend.jacfwd(f)
+    np.testing.assert_allclose(jacf(x), np.ones([5, 3]), atol=1e-5)
+
+
 def test_jax_svd(jaxb, highp):
     def l(A):
         u, _, v, _ = tc.backend.svd(A)